+++ src/EntityGraphQL/Schema/SchemaGenerator.cs
@@ -54,7 +54,7 @@ type RootQuery {{
             foreach (var item in schema.GetMutations())
             {
                 if (!string.IsNullOrEmpty(item.Description))
-                    mutations.AppendLine($"\t\"{EscapeString(item.Description)}\"");
+                    mutations.AppendLine($"\t\"\"\"{EscapeString(item.Description)}\"\"\"");
 
                 mutations.AppendLine($"\t{ToCamelCaseStartsLower(item.Name)}{GetGqlArgs(item, "")}: {item.ReturnType.GqlTypeForReturnOrArgument}");
             }
@@ -72,7 +72,7 @@ type RootQuery {{
 
                 types.AppendLine();
                 if (!string.IsNullOrEmpty(typeItem.Description))
-                    types.AppendLine($"\t\"{EscapeString(typeItem.Description)}\"");
+                    types.AppendLine($"\t\"\"\"{EscapeString(typeItem.Description)}\"\"\"");
 
                 types.AppendLine($"enum {typeItem.Name} {{");
                 foreach (var field in typeItem.GetFields())
@@ -81,7 +81,7 @@ type RootQuery {{
                         continue;
 
                     if (!string.IsNullOrEmpty(field.Description))
-                        types.AppendLine($"\t\"{EscapeString(field.Description)}\"");
+                        types.AppendLine($"\t\"\"\"{EscapeString(field.Description)}\"\"\"");
 
                     types.AppendLine($"\t{field.Name}");
 
@@ -102,7 +102,7 @@ type RootQuery {{
 
                 types.AppendLine();
                 if (!string.IsNullOrEmpty(typeItem.Description))
-                    types.AppendLine($"\"{EscapeString(typeItem.Description)}\"");
+                    types.AppendLine($"\"\"\"{EscapeString(typeItem.Description)}\"\"\"");
 
                 types.AppendLine($"{(typeItem.IsInput ? "input" : "type")} {typeItem.Name} {{");
                 foreach (var field in typeItem.GetFields())
@@ -111,7 +111,7 @@ type RootQuery {{
                         continue;
 
                     if (!string.IsNullOrEmpty(field.Description))
-                        types.AppendLine($"\t\"{EscapeString(field.Description)}\"");
+                        types.AppendLine($"\t\"\"\"{EscapeString(field.Description)}\"\"\"");
 
                     types.AppendLine($"\t{ToCamelCaseStartsLower(field.Name)}{GetGqlArgs(field)}: {field.ReturnType.GqlTypeForReturnOrArgument}");
                 }
@@ -141,7 +141,7 @@ type RootQuery {{
                 if (t.Name.StartsWith("__"))
                     continue;
                 if (!string.IsNullOrEmpty(t.Description))
-                    sb.AppendLine($"\t\"{EscapeString(t.Description)}\"");
+                    sb.AppendLine($"\t\"\"\"{EscapeString(t.Description)}\"\"\"");
                 sb.AppendLine($"\t{ToCamelCaseStartsLower(t.Name)}{GetGqlArgs(t)}: {t.ReturnType.GqlTypeForReturnOrArgument}");
             }